[ALPS08794652] f2fs: add debug to track who freeze f2fs
[Description]
There is an f2fs issue hanging in:
 issue_discard_thread -->
 sb_start_intwrite(sb) ->
 __sb_start_write(sb, SB_FREEZE_FS) ->
 percpu_down_read(sb->s_writers.rw_sem + level -1)

The only candidate blocker for this semaphore is located in:
 freeze_super@fs/super.c -->
 sb_wait_write(sb, SB_FREEZE_FS) ->
 percpu_down_write(sb->s_writers.rw_sem + level -1)

However, freeze_super is only invoked in 2 cases:
1. f2fs gc
2. ioctl specific for freeze filesystem

Since gc is not performed when hang occurs, so we add debug log to
check if there is any user space process using ioctl to freeze
filesystem.

[Test]
Boot to homescreen. File Copy.

Signed-off-by: Light Hsieh <light.hsieh@mediatek.com>
CR-Id: ALPS08794652
Feature: F2FS FileSystem
Change-Id: I281c29014b052344c241c0aef1d2aedd4ec8ffdb