[ALPS08961032] gzvm: fix double mapping issue
[Description]
In very low memory condition, the `get_user_pages` function may not
actually allocate a new page for us which acts somehow like copy-on-
write. And we invoke `pin_user_pages` consequently which trigger kernel
to break the copy-on-write, that is, allocating a new page.

But we use the old page to do mapping and protection, resulting in
(1) VM does not access to the correct physical pages and
(2) GZ protects the wrong pages and trigger EL1 data abort exception.

The solution is to invoke pin_user_pages firstly in order to get the
actuall physical page and make sure the physical address will not
changed or swapped.

[Test]
Microdroid Test App on k6991v1_64
Microdroid Test App on k6991v1_64 w/ low memory

Signed-off-by: Jerry Wang <ze-yu.wang@mediatek.com>
CR-Id: ALPS08961032
Feature: Android Virtualization Framework
Change-Id: Ifcfab4b1b7f56125e4cf134e858c97e79b0cb8a4